requirements.txt
appdirs==1.4.3
asn1crypto==1.2.0
astroid==2.3.2
attrs==19.1.0
bottle==0.12.17
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
chardet==3.0.4
coverage==4.5.4
Django==2.2.8
django-appconf==1.0.3
django-autoslug==1.9.6
django-compressor==2.3
django-debug-toolbar==2.0
django-haystack==2.8.1
django-ipware==2.1.0
django-mdeditor==0.1.16
django-uuslug==1.1.9
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0
idna==2.8
ipaddress==1.0.22
isort==4.3.21
jieba==0.39
jsonpickle==1.2
lazy-object-proxy==1.4.2
markdown2==2.3.8
mccabe==0.6.1
mistune==0.8.4
olefile==0.46
packaging==19.2
Pillow==6.2.1
pycparser==2.19
Pygments==2.4.2
pylint==2.4.3
pyparsing==2.4.2
python-dateutil==2.8.0
python-logstash==0.4.6
python-memcached==1.59
python-slugify==4.0.0
pytz==2019.3
raven==6.10.0
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
webencodings==0.5.1
WeRoBot==1.10.1
Whoosh==2.7.4
wrapt==1.11.1
xmltodict==0.12.0
PyMySQL==0.9.3
uWSGI==2.0.18